too much copy&paste.. should use AsDouble instead of Int64 :P

+++ NSupport.Test/StringToNumberConversionTest.cs
@@ -370,7 +370,7 @@
             };
 
             foreach (var value in invalidValues) {
-                Assert.Null(value.AsInt64());
+                Assert.Null(value.AsDouble());
             }
         }
